requirements.txt
asn1crypto==0.24.0
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-cleanup==1.0.1
django-cors-headers==2.1.0
django-debug-toolbar==1.8
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
pycparser==2.18
PyJWT==1.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
